code/requirements.txt
azure-functions
langchain
azure-search-documents
streamlit==1.20.0
openai==0.27.2
matplotlib==3.6.3
plotly==5.12.0
scipy==1.10.0
scikit-learn==1.2.0
transformers==4.25.1
redis==4.5.4
python-dotenv==1.0.0
azure-ai-formrecognizer==3.2.0
azure-storage-blob==12.14.1
requests==2.28.2
tiktoken==0.2.0
azure-storage-queue==12.5.0
# langchain==0.0.136
beautifulsoup4==4.12.0
streamlit-chat==0.0.2.2
fake-useragent==1.1.3
chardet==5.1.0